requirements.txt
accelerate
bitsandbytes
datasets
exrex
fire
graphviz
huggingface-hub
hydra-core
hydra-optuna-sweeper
ipykernel
ipython
jupyterlab
omegaconf
optuna
pandas
peft
psutil
regex
transformers
wandb
